ciphey/BruteForceStegpy/requirement.txt
numpy==1.23.5
Pillow==9.4.0
cryptography==37.0.0
opencv-python==4.4.0.44
opencv-python-headless==4.5.4.60